Big refactor of logic into login_server.rs and client.rs
